requirements.txt
opencv-python==4.8.1.78
pytesseract==0.3.10
pillow==10.0.1
requests==2.31.0